Basic Cost Constraining
Had to do a minor refactor of the Ast. Introduced a concept of
`Program`, wherein the `EmptyProgram` and all other `Ast`s are
represented.

Added generators for various types of Ast, those that have cost
constraints specified and those that don't. And with generators that
create invalid or valid versions of those combinations.

Program validation has also been updated to ensure that maximumCost from
a `CostConstraint` is taken into account while validating the program.