Added a dsl for records
They're not quite records, but the language itself isn't so much the
point. Proves that I can use the Kotlin DSL feature to make a "parser"
and "compiler".

The naming is very mess and lots of systematizing needs to happen.
But this is an initial proof of concept that I think demonstrates that
it's viable enough to generate an IR which can ultimately output
the rest of the code I want.